c/float_asm.c: fix compilation with clang 3.8, again
Alpine clang does not like using %f/%g print format with float. The
specification indeed states that a double value is expected.

Fixes:
    float_asm.c:242:5: error: implicit conversion increases
    floating-point precision: 'const float' to 'double'
    [-Werror,-Wdouble-promotion]
        _ok_f("fldz", f, 0);
        ^~~~~~~~~~~~~~~~~~~
    float_asm.c:222:56: note: expanded from macro '_ok_f'
            _ok(description, (fabsf(r - e) < 1.e-6f), "g", r, e); \
            ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~
    float_asm.c:205:62: note: expanded from macro '_ok'
                printf("[ OK ] %s = %" format "\n", description, result); \
                ~~~~~~                                           ^~~~~~

+++ c/float_asm.c
@@ -219,7 +219,7 @@ static double invsqrt_d(double x)
 #define _ok_f(description, result, expected) \
     do { \
         const float r = (result), e = (expected); \
-        _ok(description, (fabsf(r - e) < 1.e-6f), "g", r, e); \
+        _ok(description, (fabsf(r - e) < 1.e-6f), "g", (double)r, (double)e); \
     } while (0)
 #define ok_f(fctcall, value) _ok_f(#fctcall, (fctcall), (value))
 
@@ -230,7 +230,7 @@ static double invsqrt_d(double x)
     } while (0)
 #define ok_d(fctcall, value) _ok_d(#fctcall, (fctcall), (value))
 
-#define ok_nan(fctcall) _ok(#fctcall, isnan(fctcall), "f", (fctcall), 0./0.)
+#define ok_nan(fctcall) _ok(#fctcall, isnan(fctcall), "f", (double)(fctcall), 0./0.)
 
 static int check_constants(void)
 {
